// src/gen/ts/service-all_grpc.ts
type ClientMethodDefinition<RequestType, ResponseType> = Omit<MethodDefinition<RequestType, ResponseType>, "requestDeserialize" | "responseSerialize">;

/**
 * @generated from service service.ServiceAll
 */
export const ServiceAllDefinition: {
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary: ClientMethodDefinition<MessageInitShape<typeof StringValueSchema>, Int32Value> & ServerMethodDefinition<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ServerStream
   */
  serverStream: ClientMethodDefinition<MessageInitShape<typeof StringValueSchema>, Int32Value> & ServerMethodDefinition<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream: ClientMethodDefinition<MessageInitShape<typeof StringValueSchema>, Int32Value> & ServerMethodDefinition<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.Bidi
   * @deprecated
   */
  bidi: ClientMethodDefinition<MessageInitShape<typeof StringValueSchema>, Int32Value> & ServerMethodDefinition<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
} = {
  unary: {
    path: "/service.ServiceAll/Unary",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: MessageInitShape<typeof StringValueSchema>): Buffer => Buffer.from(toBinary(StringValueSchema, create(StringValueSchema, value))),
    requestDeserialize: (bytes: Buffer): StringValue => fromBinary(StringValueSchema, bytes),
    responseSerialize: (value: MessageInitShape<typeof Int32ValueSchema>): Buffer => Buffer.from(toBinary(Int32ValueSchema, create(Int32ValueSchema, value))),
    responseDeserialize: (bytes: Buffer): Int32Value => fromBinary(Int32ValueSchema, bytes),
  },
  serverStream: {
    path: "/service.ServiceAll/ServerStream",
    requestStream: false,
    responseStream: true,
    requestSerialize: (value: MessageInitShape<typeof StringValueSchema>): Buffer => Buffer.from(toBinary(StringValueSchema, create(StringValueSchema, value))),
    requestDeserialize: (bytes: Buffer): StringValue => fromBinary(StringValueSchema, bytes),
    responseSerialize: (value: MessageInitShape<typeof Int32ValueSchema>): Buffer => Buffer.from(toBinary(Int32ValueSchema, create(Int32ValueSchema, value))),
    responseDeserialize: (bytes: Buffer): Int32Value => fromBinary(Int32ValueSchema, bytes),
  },
  clientStream: {
    path: "/service.ServiceAll/ClientStream",
    requestStream: true,
    responseStream: false,
    requestSerialize: (value: MessageInitShape<typeof StringValueSchema>): Buffer => Buffer.from(toBinary(StringValueSchema, create(StringValueSchema, value))),
    requestDeserialize: (bytes: Buffer): StringValue => fromBinary(StringValueSchema, bytes),
    responseSerialize: (value: MessageInitShape<typeof Int32ValueSchema>): Buffer => Buffer.from(toBinary(Int32ValueSchema, create(Int32ValueSchema, value))),
    responseDeserialize: (bytes: Buffer): Int32Value => fromBinary(Int32ValueSchema, bytes),
  },
  bidi: {
    path: "/service.ServiceAll/Bidi",
    requestStream: true,
    responseStream: true,
    requestSerialize: (value: MessageInitShape<typeof StringValueSchema>): Buffer => Buffer.from(toBinary(StringValueSchema, create(StringValueSchema, value))),
    requestDeserialize: (bytes: Buffer): StringValue => fromBinary(StringValueSchema, bytes),
    responseSerialize: (value: MessageInitShape<typeof Int32ValueSchema>): Buffer => Buffer.from(toBinary(Int32ValueSchema, create(Int32ValueSchema, value))),
    responseDeserialize: (bytes: Buffer): Int32Value => fromBinary(Int32ValueSchema, bytes),
  },
};

/**
 * @generated from service service.ServiceAll
 */
export interface ServiceAllServer extends UntypedServiceImplementation {
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary: handleUnaryCall<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ServerStream
   */
  serverStream: handleServerStreamingCall<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream: handleClientStreamingCall<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.Bidi
   * @deprecated
   */
  bidi: handleBidiStreamingCall<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
}

/**
 * @generated from service service.ServiceAll
 */
export class ServiceAllClient extends Client {
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary(request: MessageInitShape<typeof StringValueSchema>, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary(request: MessageInitShape<typeof StringValueSchema>, metadata: Metadata, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary(request: MessageInitShape<typeof StringValueSchema>, options: CallOptions, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary(request: MessageInitShape<typeof StringValueSchema>, metadata: Metadata, options: CallOptions, callback: requestCallback<Int32Value>): ClientUnaryCall;
  unary(request: MessageInitShape<typeof StringValueSchema>, metadataOrOptionsOrCallback?: Metadata | CallOptions | requestCallback<Int32Value>, optionsOrCallback?: CallOptions | requestCallback<Int32Value>, callback?: requestCallback<Int32Value>): ClientUnaryCall {
    return this.makeUnaryRequest(
      ServiceAllDefinition.unary.path,
      ServiceAllDefinition.unary.requestSerialize,
      ServiceAllDefinition.unary.responseDeserialize,
      request,
      metadataOrOptionsOrCallback as Metadata,
      optionsOrCallback as CallOptions,
      callback as requestCallback<Int32Value>,
    );
  }
  /**
   * @generated from rpc service.ServiceAll.ServerStream
   */
  serverStream(request: MessageInitShape<typeof StringValueSchema>, metadata?: Metadata, options?: CallOptions): ClientReadableStream<Int32Value>;
  /**
   * @generated from rpc service.ServiceAll.ServerStream
   */
  serverStream(request: MessageInitShape<typeof StringValueSchema>, options?: CallOptions): ClientReadableStream<Int32Value>;
  serverStream(request: MessageInitShape<typeof StringValueSchema>, metadataOrOptions?: Metadata | CallOptions, options?: CallOptions): ClientReadableStream<Int32Value> {
    return this.makeServerStreamRequest(
      ServiceAllDefinition.serverStream.path,
      ServiceAllDefinition.serverStream.requestSerialize,
      ServiceAllDefinition.serverStream.responseDeserialize,
      request,
      metadataOrOptions as Metadata,
      options,
    );
  }
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream(callback: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream(metadata: Metadata, callback: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream(options: CallOptions, callback: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream(metadata: Metadata, options: CallOptions, callback: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>>;
  clientStream(metadataOrOptionsOrCallback?: Metadata | CallOptions | requestCallback<Int32Value>, optionsOrCallback?: CallOptions | requestCallback<Int32Value>, callback?: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>> {
    return this.makeClientStreamRequest(
      ServiceAllDefinition.clientStream.path,
      ServiceAllDefinition.clientStream.requestSerialize,
      ServiceAllDefinition.clientStream.responseDeserialize,
      metadataOrOptionsOrCallback as Metadata,
      optionsOrCallback as CallOptions,
      callback as requestCallback<Int32Value>,
    );
  }
  /**
   * @generated from rpc service.ServiceAll.Bidi
   * @deprecated
   */
  bidi(metadata?: Metadata, options?: CallOptions): ClientDuplexStream<MessageInitShape<typeof StringValueSchema>, Int32Value>;
  /**
   * @generated from rpc service.ServiceAll.Bidi
   * @deprecated
   */
  bidi(options?: CallOptions): ClientDuplexStream<MessageInitShape<typeof StringValueSchema>, Int32Value>;
  bidi(metadataOrOptions?: Metadata | CallOptions, options?: CallOptions): ClientDuplexStream<MessageInitShape<typeof StringValueSchema>, Int32Value> {
    return this.makeBidiStreamRequest(
      ServiceAllDefinition.bidi.path,
      ServiceAllDefinition.bidi.requestSerialize,
      ServiceAllDefinition.bidi.responseDeserialize,
      metadataOrOptions as Metadata,
      options,
    );
  }
}