// src/gen/js/service-all_grpc.d.ts
type ClientMethodDefinition<RequestType, ResponseType> = Omit<MethodDefinition<RequestType, ResponseType>, "requestDeserialize" | "responseSerialize">;

/**
 * @generated from service service.ServiceAll
 */
export declare const ServiceAllDefinition: {
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary: ClientMethodDefinition<MessageInitShape<typeof StringValueSchema>, Int32Value> & ServerMethodDefinition<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ServerStream
   */
  serverStream: ClientMethodDefinition<MessageInitShape<typeof StringValueSchema>, Int32Value> & ServerMethodDefinition<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream: ClientMethodDefinition<MessageInitShape<typeof StringValueSchema>, Int32Value> & ServerMethodDefinition<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.Bidi
   * @deprecated
   */
  bidi: ClientMethodDefinition<MessageInitShape<typeof StringValueSchema>, Int32Value> & ServerMethodDefinition<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
};

/**
 * @generated from service service.ServiceAll
 */
export declare interface ServiceAllServer extends UntypedServiceImplementation {
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary: handleUnaryCall<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ServerStream
   */
  serverStream: handleServerStreamingCall<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream: handleClientStreamingCall<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.Bidi
   * @deprecated
   */
  bidi: handleBidiStreamingCall<StringValue, MessageInitShape<typeof Int32ValueSchema>>;
}

/**
 * @generated from service service.ServiceAll
 */
export declare class ServiceAllClient extends Client {
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary(request: MessageInitShape<typeof StringValueSchema>, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary(request: MessageInitShape<typeof StringValueSchema>, metadata: Metadata, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary(request: MessageInitShape<typeof StringValueSchema>, options: CallOptions, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc service.ServiceAll.Unary
   */
  unary(request: MessageInitShape<typeof StringValueSchema>, metadata: Metadata, options: CallOptions, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc service.ServiceAll.ServerStream
   */
  serverStream(request: MessageInitShape<typeof StringValueSchema>, metadata?: Metadata, options?: CallOptions): ClientReadableStream<Int32Value>;
  /**
   * @generated from rpc service.ServiceAll.ServerStream
   */
  serverStream(request: MessageInitShape<typeof StringValueSchema>, options?: CallOptions): ClientReadableStream<Int32Value>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream(callback: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream(metadata: Metadata, callback: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream(options: CallOptions, callback: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.ClientStream
   */
  clientStream(metadata: Metadata, options: CallOptions, callback: requestCallback<Int32Value>): ClientWritableStream<MessageInitShape<typeof StringValueSchema>>;
  /**
   * @generated from rpc service.ServiceAll.Bidi
   * @deprecated
   */
  bidi(metadata?: Metadata, options?: CallOptions): ClientDuplexStream<MessageInitShape<typeof StringValueSchema>, Int32Value>;
  /**
   * @generated from rpc service.ServiceAll.Bidi
   * @deprecated
   */
  bidi(options?: CallOptions): ClientDuplexStream<MessageInitShape<typeof StringValueSchema>, Int32Value>;
}