// src/gen/ts/service-edge-cases_grpc.ts
type ClientMethodDefinition<RequestType, ResponseType> = Omit<MethodDefinition<RequestType, ResponseType>, "requestDeserialize" | "responseSerialize">;

/**
 * A service without any methods.
 *
 * @generated from service service.ServiceEmpty
 */
export const ServiceEmptyDefinition: Record<string, never> = {
};

/**
 * A service without any methods.
 *
 * @generated from service service.ServiceEmpty
 */
export interface ServiceEmptyServer extends UntypedServiceImplementation {
}

/**
 * A service without any methods.
 *
 * @generated from service service.ServiceEmpty
 */
export class ServiceEmptyClient extends Client {
}

/**
 * A service with method names that require escaping in generated code.
 *
 * @generated from service service.ServiceNames
 */
export const ServiceNamesDefinition: {
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$: ClientMethodDefinition<MessageInitShape<typeof EmptySchema>, Empty> & ServerMethodDefinition<Empty, MessageInitShape<typeof EmptySchema>>;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete: ClientMethodDefinition<MessageInitShape<typeof EmptySchema>, Empty> & ServerMethodDefinition<Empty, MessageInitShape<typeof EmptySchema>>;
} = {
  constructor$: {
    path: "/service.ServiceNames/Constructor",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: MessageInitShape<typeof EmptySchema>): Buffer => Buffer.from(toBinary(EmptySchema, create(EmptySchema, value))),
    requestDeserialize: (bytes: Buffer): Empty => fromBinary(EmptySchema, bytes),
    responseSerialize: (value: MessageInitShape<typeof EmptySchema>): Buffer => Buffer.from(toBinary(EmptySchema, create(EmptySchema, value))),
    responseDeserialize: (bytes: Buffer): Empty => fromBinary(EmptySchema, bytes),
  },
  delete: {
    path: "/service.ServiceNames/Delete",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: MessageInitShape<typeof EmptySchema>): Buffer => Buffer.from(toBinary(EmptySchema, create(EmptySchema, value))),
    requestDeserialize: (bytes: Buffer): Empty => fromBinary(EmptySchema, bytes),
    responseSerialize: (value: MessageInitShape<typeof EmptySchema>): Buffer => Buffer.from(toBinary(EmptySchema, create(EmptySchema, value))),
    responseDeserialize: (bytes: Buffer): Empty => fromBinary(EmptySchema, bytes),
  },
};
// ...
export class ServiceNamesClient extends Client {
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$(request: MessageInitShape<typeof EmptySchema>, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$(request: MessageInitShape<typeof EmptySchema>, metadata: Metadata, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$(request: MessageInitShape<typeof EmptySchema>, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$(request: MessageInitShape<typeof EmptySchema>, metadata: Metadata, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  constructor$(request: MessageInitShape<typeof EmptySchema>, metadataOrOptionsOrCallback?: Metadata | CallOptions | requestCallback<Empty>, optionsOrCallback?: CallOptions | requestCallback<Empty>, callback?: requestCallback<Empty>): ClientUnaryCall {
    return this.makeUnaryRequest(
      ServiceNamesDefinition.constructor$.path,
      ServiceNamesDefinition.constructor$.requestSerialize,
      ServiceNamesDefinition.constructor$.responseDeserialize,
      request,
      metadataOrOptionsOrCallback as Metadata,
      optionsOrCallback as CallOptions,
      callback as requestCallback<Empty>,
    );
  }
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete(request: MessageInitShape<typeof EmptySchema>, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete(request: MessageInitShape<typeof EmptySchema>, metadata: Metadata, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete(request: MessageInitShape<typeof EmptySchema>, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete(request: MessageInitShape<typeof EmptySchema>, metadata: Metadata, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  delete(request: MessageInitShape<typeof EmptySchema>, metadataOrOptionsOrCallback?: Metadata | CallOptions | requestCallback<Empty>, optionsOrCallback?: CallOptions | requestCallback<Empty>, callback?: requestCallback<Empty>): ClientUnaryCall {
    return this.makeUnaryRequest(
      ServiceNamesDefinition.delete.path,
      ServiceNamesDefinition.delete.requestSerialize,
      ServiceNamesDefinition.delete.responseDeserialize,
      request,
      metadataOrOptionsOrCallback as Metadata,
      optionsOrCallback as CallOptions,
      callback as requestCallback<Empty>,
    );
  }
}