// src/gen/js/service-edge-cases_grpc.d.ts
export declare class ServiceNamesClient extends Client {
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$(request: MessageInitShape<typeof EmptySchema>, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$(request: MessageInitShape<typeof EmptySchema>, metadata: Metadata, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$(request: MessageInitShape<typeof EmptySchema>, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * The local name "constructor" cannot be used as a class method.
   *
   * @generated from rpc service.ServiceNames.Constructor
   */
  constructor$(request: MessageInitShape<typeof EmptySchema>, metadata: Metadata, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete(request: MessageInitShape<typeof EmptySchema>, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete(request: MessageInitShape<typeof EmptySchema>, metadata: Metadata, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete(request: MessageInitShape<typeof EmptySchema>, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
  /**
   * Reserved words are valid class method and object property names.
   *
   * @generated from rpc service.ServiceNames.Delete
   */
  delete(request: MessageInitShape<typeof EmptySchema>, metadata: Metadata, options: CallOptions, callback: requestCallback<Empty>): ClientUnaryCall;
}